+++ lib/modules/network/network_service.dart
@@ -1,9 +1,9 @@
 ﻿// Module: network
-// Updated at: 2026-05-27 10:45:51
-// Commit: chore(network): update dependencies and cleanup unused imports in network
+// Updated at: 2026-05-27 11:21:17
+// Commit: style(network): harmonize light and dark theme palette for network
 
 class networkComponentState {
-  final String commitRef = '1484';
-  final String lastMessage = 'chore(network): update dependencies and cleanup unused imports in network';
-  final DateTime timestamp = DateTime.parse('2026-05-27T10:45:51+05:00');
+  final String commitRef = '1487';
+  final String lastMessage = 'style(network): harmonize light and dark theme palette for network';
+  final DateTime timestamp = DateTime.parse('2026-05-27T11:21:17+05:00');
 }